requirements.txt
flask
paho-mqtt
cryptography
flask_socketio